Release-13.30.0

+++ app/controllers/batches/comments_controller.rb
@@ -16,9 +16,7 @@ class Batches::CommentsController < ApplicationController # rubocop:todo Style/D
 
   def destroy
     comment = Comment.find(params[:id])
-    if comment.present?
-      comment.destroy
-    end
+    comment.destroy if comment.present?
     @comments = @batch.comments
     render partial: 'list', locals: { commentable: @batch, visible: true }
   end
